requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp==3.9.5
websockets>=11.0.0
cryptography==42.0.8
python-socketio[asyncio_client]==5.11.2
python-engineio==4.9.1

# Basic utilities
colorama>=0.4.6
tabulate>=0.9.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
